Changed how autonomous works, in preparation for adding the modular functionality
	Firstly, the autonomous commands have been changed to wpilib Commands,
so we don't have to worry about scheduling our own commands, we can just
start our desired command. I also created a DoNothing Command, as there
was not one previously.
	
	IMPORTANT THING: I tried to migrate the commands as best I could, but
they still really need to be reviewed and tested, as they may not work
the same as before (They still should, I might have just failed at
converting them)
	
	
	All of the autonomous commands are added to the smart dashboard as a
sendable chooser, and during autonomousInit(), the autoCommand object is
set to be the currently selected option. autoCommand is then scheduled
by the schedule by calling it's start method.
	
	No more EnumAutoCMD and AutonomousCommand classes, which were used to
do our own command running with the "doCMD()" and "done()" methods, we
are going to use wpilib's scheduler to automate running the commands.
This is necessary because I plan to use wpilib's CommandGroup class to
run the "modules" of modular autonomous mode
	
	No more ICommandOption, for now, the rewrite deprecated it, and it
wasn't used at all, so I deleted it. It may be brought back so that the
autonomous commands can be referred to polymorphically, but that's for
later
	
SUPER IMPORTANT: Seriously, the rewritten commands need to be tested to
make sure they work the same as before

DOUBLE SUPER IMPORTANT: see above